runner: remaining defensive guards (RN15); RN14 already guarded
RN14: confirmed model.c already refuses q8 KV and falls back to f16 when any
layer's head_dim % 32 != 0 (guard fires in the test log) — no change.

RN15 guards:
- engine.c: bound the lp_capture_pre snapshot copy to the snap[] size.
- model.c: check osc->ne[0] >= 1 before the layer_output_scale scalar read.
- runner.h: widen sframe.idx int16_t -> int32_t so a large maxItems can't wrap
  negative and silently drop enforcement (raises the ceiling to INT_MAX,
  matching compile_bound; keeps the huge-min-bounds terminate test green).
- server.c: free a partially-initialized tokenizer on swap-load failure.
- cuda.c: NULL-guard g = m->gpu in gpu_forward_batch (post gpu_disable recovery);
  free already-created events on prof_init mid-loop failure.
Documented (no code change): the dev-only global prof race, and the power-of-two
blockDim launch invariants in kernels.cu (hold by construction). Skipped
keep_alive->global TTL (intentional Ollama-style single-resident-model behavior;
scoping it needs a protocol/restructure change, out of scope).

Built clean; make test green; TinyLlama GPU load+generate smoke ok. PTX unchanged.

+++ src/runner.h
@@ -666,7 +666,9 @@ void   schema_free(snode *n);
 typedef struct {
     const snode *node;
     uint8_t  phase, sub;
-    int16_t  idx;
+    int32_t  idx;       // array element / object key counter (int32: max/minItems
+                        // are bounded only by INT_MAX, so int16 could wrap and
+                        // silently drop maxItems enforcement past 32767)
     int32_t  lit_pos;   // string chars / literal bytes seen (int32: file-sized strings)
     uint16_t disc;      // this object's discriminator choice + 1; 0 = not chosen yet
     uint64_t alive;
